Meshes can now be loaded from .obj files
